+++ spec/stack_master/commands/delete_spec.rb
@@ -1,7 +1,7 @@
 RSpec.describe StackMaster::Commands::Delete do
 
   subject(:delete) { described_class.new(stack_name, region) }
-  let(:cf) { Aws::CloudFormation::Client.new }
+  let(:cf) { spy(Aws::CloudFormation::Client.new) }
   let(:region) { 'us-east-1' }
   let(:stack_name) { 'mystack' }
 
@@ -15,24 +15,26 @@ RSpec.describe StackMaster::Commands::Delete do
   describe "#perform" do
     context "The stack exists" do
       before do
-        cf.stub_responses(:describe_stacks, stacks: [{ stack_id: "ABC", stack_name: stack_name, creation_time: Time.now, stack_status: 'UPDATE_COMPLETE', parameters: []}])
-
+        allow(cf).to receive(:describe_stacks).and_return(
+          {stacks: [{ stack_id: "ABC", stack_name: stack_name, creation_time: Time.now, stack_status: 'UPDATE_COMPLETE', parameters: []}]}
+        )
       end
       it "deletes the stack and tails the events" do
-        expect(cf).to receive(:delete_stack).with({:stack_name => region})
-        expect(StackMaster::StackEvents::Streamer).to receive(:stream)
         delete.perform
+        expect(cf).to have_received(:delete_stack).with({:stack_name => region})
+        expect(StackMaster::StackEvents::Streamer).to have_received(:stream)
       end
     end
 
     context "The stack does not exist" do
       before do
-        cf.stub_responses(:describe_stacks, Aws::CloudFormation::Errors::ValidationError.new("x", "y"))
+        allow(cf).to receive(:describe_stacks).and_raise(Aws::CloudFormation::Errors::ValidationError.new("x", "y"))
       end
-      it "raises an error" do
-        expect(StackMaster::StackEvents::Streamer).to_not receive(:stream)
-        expect(cf).to_not receive(:delete_stack)
+      it "is not successful" do
         delete.perform
+        expect(StackMaster::StackEvents::Streamer).not_to have_received(:stream)
+        expect(cf).not_to have_received(:delete_stack)
+        expect(delete.success?).to be false
       end
     end
   end